fix(just): use --tag flag in changelog generation for correct version
The 'just tag' command was calling 'just changelog' which used
'git cliff --latest' without specifying the new version tag.
This caused git-cliff to generate changelog entries under the
previous version instead of the new one.

Now we pass --tag "v$VERSION" directly to git-cliff so it
knows which version to generate the changelog for, even before
the git tag is created.